Completed bearer context setup cycle To be tested: Handling of bearer context setup response and subsequent UE context modifition msg to F1AP task

+++ openair2/COMMON/e1ap_messages_types.h
@@ -121,7 +121,7 @@ typedef struct DRB_nGRAN_to_setup_s {
 typedef struct pdu_session_to_setup_s {
   long sessionId;
   long sessionType;
-  int32_t sst;
+  int8_t sst;
   long integrityProtectionIndication;
   long confidentialityProtectionIndication;
   in_addr_t tlAddress;
